requirements.txt
requests
pandas
sqlalchemy
pymysql
tqdm
matplotlib
seaborn
plotly
streamlit
mysql-connector-python
python-dotenv